refactor(wallet): move types from get balances to wallet types
