+++ libs/solana-dapp/wallet/types/src/lib/types.ts
@@ -41,3 +41,15 @@ export interface Wallet {
 }
 
 export const DEFAULT_WALLET = WalletName.Sollet;
+
+export interface Balance {
+  address: string;
+  lamports: number;
+  quantity: number;
+  hasBalance: boolean;
+}
+
+export interface GetBalancesConfig {
+  rpcEndpoint: string;
+  walletPublicKey: string;
+}
