+++ libs/solana-dapp/wallet/utils/get-balances/src/index.ts
@@ -1,2 +1,1 @@
 export * from './lib/get-balances';
-export * from './lib/types';
