+++ libs/solana-dapp/wallet/utils/get-balances/src/lib/get-balances.ts
@@ -1,9 +1,9 @@
+import { Balance, GetBalancesConfig } from '@nx-dapp/solana-dapp/wallet/types';
 import { Connection, PublicKey } from '@solana/web3.js';
 import { Observable, of } from 'rxjs';
 import { switchMap } from 'rxjs/operators';
 
 import { mapToBalances, observeUserAccounts } from './operators';
-import { Balance, GetBalancesConfig } from './types';
 import { getMintAccounts, getUserAccounts } from './utils';
 
 export const getBalances = (config: GetBalancesConfig): Observable<Balance[]> =>
