--- libs/solana-dapp/wallet/utils/get-balances/src/lib/types.ts
@@ -1,11 +1,0 @@
-export interface Balance {
-  address: string;
-  lamports: number;
-  quantity: number;
-  hasBalance: boolean;
-}
-
-export interface GetBalancesConfig {
-  rpcEndpoint: string;
-  walletPublicKey: string;
-}
